requirements.txt
groq
datasets
pandas
python-dotenv